requirements.txt
langchain
langchain-openai
langchain-community
langgraph
openai
duckdb
duckdb-engine
pandas
matplotlib
python-dotenv
streamlit
